+++ remote/src/main/java/com/junjange/remote/api/ApiService.kt
@@ -1,18 +1,22 @@
 package com.junjange.remote.api
 
 import com.junjange.remote.model.BaseResponse
+import com.junjange.remote.model.request.LotteryRandomRequest
+import com.junjange.remote.model.request.PensionLotteryRandomRequest
 import com.junjange.remote.model.request.RefreshRequest
 import com.junjange.remote.model.request.RegisterRequest
-import com.junjange.remote.model.response.AccessTokenResponse
 import com.junjange.remote.model.response.IsRegisteredResponse
 import com.junjange.remote.model.response.JwtTokenResponse
+import com.junjange.remote.model.response.LotteryRandomResponse
+import com.junjange.remote.model.response.LotteryResultResponse
+import com.junjange.remote.model.response.PensionLotteryRandomResponse
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.Query
 
 
-internal interface CredentialApiService {
+internal interface ApiService {
 
     @POST(ApiClient.Credentials.postCredentialsRegister)
     suspend fun postRegister(
@@ -37,5 +41,24 @@ internal interface CredentialApiService {
         @Query("idToken") idToken: String,
         @Query("provider") provider: String,
     ): BaseResponse<IsRegisteredResponse>
+
+    @POST(ApiClient.Lottery.postLotterySave)
+    suspend fun postLotterySave(
+        @Body body: LotteryRandomRequest,
+    ): BaseResponse<Unit>
+
+    @GET(ApiClient.Lottery.getLotteryRandom)
+    suspend fun getLotteryRandom(): BaseResponse<LotteryRandomResponse>
+
+    @GET(ApiClient.Lottery.getLotteryGet)
+    suspend fun getLotteryGet(
+        @Body body: RefreshRequest,
+    ): BaseResponse<LotteryResultResponse>
+
+    @POST(ApiClient.PensionLottery.postPensionLotterySave)
+    suspend fun postPensionLotterySave(@Body body: PensionLotteryRandomRequest): BaseResponse<Unit>
+
+    @GET(ApiClient.PensionLottery.getPensionLotteryRandom)
+    suspend fun getPensionLotteryRandom(): BaseResponse<PensionLotteryRandomResponse>
 
 }
