+++ remote/src/main/java/com/junjange/remote/datasource/LotteryDataSourceImpl.kt
@@ -2,13 +2,13 @@ package com.junjange.remote.datasource
 
 import com.junjange.data.datasource.LotteryDataSource
 import com.junjange.data.model.remote.LotteryRandomEntity
-import com.junjange.remote.api.LotteryApiService
+import com.junjange.remote.api.ApiService
 import com.junjange.remote.model.request.LotteryRandomRequest
 import com.junjange.remote.model.response.toData
 import javax.inject.Inject
 
 internal class LotteryDataSourceImpl @Inject constructor(
-    private val apiService: LotteryApiService,
+    private val apiService: ApiService,
 ) : LotteryDataSource {
 
     override suspend fun postLotterySave(
